✅ test : 루프 충돌 해결(pubsub루프가 충돌되어서 난 에러였음)
